+++ src/app/Message/MessageValidatorInterface.php
@@ -1,0 +1,18 @@
+<?php
+
+declare(strict_types=1);
+
+namespace AuthorsDMS\Message;
+
+interface MessageValidatorInterface
+{
+    /**
+     * @return bool
+     */
+    public function isValid(): bool;
+
+    /**
+     * @return array
+     */
+    public function getErrors(): array;
+}