+++ src/app/Providers/MessageRoutingServiceProvider.php
@@ -4,14 +4,12 @@ declare(strict_types=1);
 
 namespace AuthorsDMS\Providers;
 
+use AuthorsDMS\OutboundAdapters\AuthorCreatedEvent;
+use AuthorsDMS\OutboundAdapters\AuthorUpdatedEvent;
+use AuthorsDMS\Services\PostEventsService;
 use Chassis\Framework\Providers\RoutingServiceProvider;
-use AuthorsDMS\OutboundAdapters\DemoOperationDelete;
-use AuthorsDMS\OutboundAdapters\DemoOperationDeletedEvents;
-use AuthorsDMS\OutboundAdapters\DemoOperationGetAsync;
-use AuthorsDMS\OutboundAdapters\DemoOperationGetSync;
-use AuthorsDMS\Services\DemoDeleteEventService;
-use AuthorsDMS\Services\DemoMoreService;
-use AuthorsDMS\Services\DemoService;
+use AuthorsDMS\OutboundAdapters\AuthorDeletedEvent;
+use AuthorsDMS\Services\AuthorService;
 
 class MessageRoutingServiceProvider extends RoutingServiceProvider
 {
@@ -19,20 +17,25 @@ class MessageRoutingServiceProvider extends RoutingServiceProvider
      * @var array|string[]
      */
     protected array $inboundRoutes = [
-        'createSomething' => [DemoService::class, 'create'],
-        'getSomething' => [DemoService::class, 'get'],
-        'getSomethingResponse' => [DemoMoreService::class, 'complete'],
-        'deleteSomething' => [DemoService::class, 'delete'],
-        'somethingDeleted' => DemoDeleteEventService::class,
+        // commands
+        'getAuthor' => [AuthorService::class, 'get'],
+        'createAuthor' => [AuthorService::class, 'create'],
+        'updateAuthor' => [AuthorService::class, 'update'],
+        'deleteAuthor' => [AuthorService::class, 'delete'],
+
+        // events
+        'postCreated' => PostEventsService::class,
+        'postUpdated' => PostEventsService::class,
+        'postDeleted' => PostEventsService::class,
     ];
 
     /**
      * @var array|string[]
      */
     protected array $outboundRoutes = [
-        'getSomethingSync' => DemoOperationGetSync::class,
-        'getSomethingAsync' => DemoOperationGetAsync::class,
-        'deleteSomething' => DemoOperationDelete::class,
-        'deleteSomethingEvent' => DemoOperationDeletedEvents::class,
+        // events
+        'authorCreated' => AuthorCreatedEvent::class,
+        'authorUpdated' => AuthorUpdatedEvent::class,
+        'authorDeleted' => AuthorDeletedEvent::class,
     ];
 }