+++ CHANGELOG.md
@@ -6,6 +6,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2022-04-08
+
+### Added
+
+- Add validator and rules per operation
+- Complete author handlers
+- Create outbound adapters
+    - events to posts microservices
+    - redis repository
+
 ## [0.1.0] - 2022-04-03
 
 ### Added